Internship-report
This project involves the physical installation of a local network with different types of cables, as well as the configuration of various network services such as DHCP, DNS, and HTTP. Additionally, the project will include the installation of network equipment such as a network printer and an IP camera. The project will conclude with the creation of a web page to test the functionality of the DHCP and DNS HTTP services.